feat: add auto pass stage for the holder account
